pyproject.toml
[tool.poetry]
name = "cava"
version = "0.1.0"
description = ""
authors = ["David Snider <david@davidsnider.org>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
pika = "^1.2.0"
fastapi = "^0.68.0"
statsd = "^3.3.0"
arrow = "^1.1.1"
rule-engine = "^3.3.0"

requests = "^2.26.0"
amcrest = "^1.8.0"
uvicorn = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
bandit = "^1.7.0"
black = "^21.6b0"
flake8 = "^3.9.2"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
